requirements.txt
langchain
langchain-openai
python-dotenv
langchain_community
langchain_core
streamlit
langchain_chroma
langchain_huggingface
langchain_groq
sentence-transformers
pypdf
arxiv
wikipedia
faiss-cpu
duckduckgo-search
sqlite3